Move get_manubot_user_agent to util & add email
